+++ src/providers/twitch/eventsub/Connection.cpp
@@ -9,6 +9,7 @@
 #include "providers/twitch/eventsub/Controller.hpp"
 #include "providers/twitch/eventsub/MessageBuilder.hpp"
 #include "providers/twitch/eventsub/MessageHandlers.hpp"
+#include "providers/twitch/PubSubManager.hpp"
 #include "providers/twitch/TwitchBadge.hpp"
 #include "providers/twitch/TwitchChannel.hpp"
 #include "providers/twitch/TwitchIrcServer.hpp"
@@ -416,12 +417,42 @@ bool Connection::isSubscribedTo(const SubscriptionRequest &request) const
 
 void Connection::markRequestSubscribed(const SubscriptionRequest &request)
 {
+    assert((this->twitchUserID.isEmpty() ||
+            this->twitchUserID == request.ownerTwitchUserID) &&
+           "A subscription was made when another user's subscriptions were "
+           "still active");
+
+    this->twitchUserID = request.ownerTwitchUserID;
+
     this->subscriptions.emplace(request);
 }
 
 void Connection::markRequestUnsubscribed(const SubscriptionRequest &request)
 {
     this->subscriptions.erase(request);
+
+    if (this->subscriptions.empty())
+    {
+        // TODO: Verify that it's fine for us to reuse a connection for another
+        // user after all old subscriptions are gone
+        this->twitchUserID.clear();
+    }
+}
+
+bool Connection::canHandleSubscriptionFrom(
+    const QString &otherTwitchUserID) const
+{
+    return this->twitchUserID.isEmpty() ||
+           this->twitchUserID == otherTwitchUserID;
+}
+
+void Connection::debug()
+{
+    for (const auto &request : this->subscriptions)
+    {
+        qCInfo(LOG).noquote().nospace()
+            << this->getSessionID() << " -> " << request;
+    }
 }
 
 }  // namespace chatterino::eventsub